+++ management/daemon.py
@@ -40,10 +40,10 @@ def authorized_personnel_only(viewfunc):
 		error = None
 		try:
 			email, privs = auth_service.authenticate(request, env)
-		except auth.MissingTokenError as e:
+		except totp.MissingTokenError as e:
 			privs = []
 			error = str(e)
-		except auth.BadTokenError as e:
+		except totp.BadTokenError as e:
 			# Write a line in the log recording the failed login
 			log_failed_login(request)
 
@@ -128,7 +128,7 @@ def me():
 	# Is the caller authorized?
 	try:
 		email, privs = auth_service.authenticate(request, env)
-	except auth.MissingTokenError as e:
+	except totp.MissingTokenError as e:
 		# Log the failed login
 		log_failed_login(request)
 
@@ -136,7 +136,7 @@ def me():
 			"status": "missing_token",
 			"reason": str(e),
 		})
-	except auth.BadTokenError as e:
+	except totp.BadTokenError as e:
 		# Log the failed login
 		log_failed_login(request)
